Refactoring
Separating the milestone processing from the merge processing and
enabling easier expansion for further additional processes.
